Update TypeScript configuration, fix import paths, and rename files
