+++ src/Softphone/index.ts
@@ -1,0 +1,6 @@
+import Softphone from "./Softphone";
+import { SoftphoneProvider } from "./context/SoftphoneProvider";
+import { useSoftphone } from "./hooks/useSoftphone";
+import { ContactInput } from "./types";
+
+export { Softphone, SoftphoneProvider, useSoftphone, type ContactInput };
