+++ src/app/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App.tsx";
 import "./index.css";
-import { SoftphoneProvider } from "./index.ts";
+import { SoftphoneProvider } from "../Softphone";
 
 ReactDOM.render(
   <React.StrictMode>
