--- src/index.ts
@@ -1,6 +1,0 @@
-import Softphone from "./Softphone/Softphone";
-import { SoftphoneProvider } from "./Softphone/context/SoftphoneProvider";
-import { useSoftphone } from "./Softphone/hooks/useSoftphone";
-import { ContactInput } from "./Softphone/types";
-
-export { Softphone, SoftphoneProvider, useSoftphone, type ContactInput };
