+++ vite.config.softphone.ts
@@ -1,0 +1,33 @@
+// vite.config.softphone.ts
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import path from "path";
+import dts from "vite-plugin-dts";
+
+export default defineConfig({
+  plugins: [react(), dts()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "src"),
+    },
+  },
+  envPrefix: "SOFTPHONE_",
+  build: {
+    lib: {
+      entry: path.resolve(__dirname, "src/Softphone/index.ts"),
+      name: "Softphone",
+      fileName: (format) => `index.${format}.js`,
+    },
+    rollupOptions: {
+      external: ["react", "react-dom"],
+      output: {
+        globals: {
+          react: "React",
+          "react-dom": "ReactDOM",
+        },
+      },
+    },
+    sourcemap: true,
+    emptyOutDir: true,
+  },
+});
